Zone.Identifier files are now handled automatically

+++ src/read_course_catalogue.py
@@ -3,30 +3,35 @@ import os
 from pypdf import PdfReader
 
 def write_pdf_to_text(file_path):
+    print(f"reading: {file_path}")
     reader = PdfReader(file_path)
 
-    output_file = file_path.replace({"source": "text", ".pdf": ".txt"})
-    print(output_file)
-    
-    with open(output_file, "w+") as file_to_write:
-        for page in reader.pages:
-            page_content = page.extract_text()
-            file_to_write.write(page_content + "\n")
+    output_file = file_path.translate({"source": "text", ".pdf": ".txt"})
+
+    # skip files which already exist
+    if not os.path.isfile(output_file):
+        with open(output_file, "w+") as file_to_write:
+            for page in reader.pages:
+                page_content = page.extract_text()
+                file_to_write.write(page_content + "\n")
 
     return output_file
 
 
 def convert_pdfs_to_txt(): 
-    source_dir = "data/source"
+    source_dir = "data/source/FS24"
     
-    files_written = []
+    new_files = []
     for path, dirs, files in os.walk(source_dir):
         for file in files:
+            if "Zone.Identifier" in file:
+                continue
+
             filename = path + "/" + file
             out = write_pdf_to_text(filename)
-            files_written.append(out)
-    return files_written
+            new_files.append(out)
+    return new_files
 
 if __name__ == "__main__":
-    written_files = convert_pdfs_to_txt()
-    print(written_files)
+    new_files = convert_pdfs_to_txt()
+    print(new_files)
